changed variable type

+++ 0x18-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	unsigned long int nodeCount = 0;
+	size_t nodeCount = 0;
 
 	while (h->prev)
 		h = h->prev;
